# src/app/docs/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Suno API Plus Next",
    "description": "Unofficial self-hosted Suno API gateway with account affinity, Studio workflows, Advanced Split, Fast Upload and OpenAI-compatible routes.",
    "version": "2.0.0"
  },
...
  "paths": {
...
    "/api/generate": {
      "post": {
        "summary": "Generate audio based on Prompt.",
        "description": "It will automatically fill in the lyrics.\n\n2 audio files will be generated for each request, consuming a total of 10 credits.\n\n`wait_audio` can be set to API mode:\n\n• By default, it is set to `false`, which indicates the background mode. It will only return audio task information, and you will need to call the get API to retrieve detailed audio information.\n\n• If set to `true`, it simulates synchronous mode. The API will wait for a maximum of 100s until the audio is generated, and will directly return the audio link and other information. Recommend using in GPTs and other agents.",
        "tags": [
          "default"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt",
                  "make_instrumental",
                  "wait_audio"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Prompt",
                    "example": "A popular heavy metal song about war, sung by a deep-voiced male singer, slowly and melodiously. The lyrics depict the sorrow of people after the war."
                  },
                  "make_instrumental": {
                    "type": "boolean",
                    "description": "Whether to generate instrumental music",
                    "example": "false"
                  },
                  "model": {
                    "type": "string",
                    "description": "Public Suno model ID. Defaults to suno-music (currently Suno V5.5 / chirp-fenix). Raw chirp-* identifiers remain pass-through compatible. Availability depends on the account plan and upstream permissions.",
                    "default": "suno-music",
                    "example": "suno-music"
                  },
                  "voice_id": {
                    "type": "string",
                    "description": "Optional public Suno +Voice UUID or https://suno.com/voice/<id> URL. Private Voices and make_instrumental=true are rejected."
                  },
                  "account_affinity": {
                    "type": "string",
                    "description": "Opaque signed value selecting the generation target account. A public Voice may be owned by a different account."
                  },
                  "persona_id": {
                    "type": "string",
                    "deprecated": true,
                    "description": "Compatibility alias for voice_id."
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Whether to wait for music generation, default is false, directly return audio task information; set to true, will wait for up to 100s until the audio is generated.",
                    "example": "false"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "0",
                      "1"
                    ],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio_info"
                      },
                      {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    ]
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/ConcurrencyLimitExceeded"
          }
        }
      }
    },
...
    "/api/cover": {
      "post": {
        "summary": "Generate a Cover from a source clip",
        "description": "Generates two Cover clips. Any source created by /api/upload_audio (including portable Studio fast uploads) must use the returned account_affinity; cross-account Cover is not supported upstream.",
        "tags": [
          "default"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "clip_id",
                  "tags",
                  "title"
                ],
                "properties": {
                  "clip_id": {
                    "type": "string",
                    "description": "Source clip returned by /api/upload_audio."
                  },
                  "account_affinity": {
                    "type": "string",
                    "description": "Owning account affinity returned by /api/upload_audio. Required for uploaded sources; do not replace it with another account. A public Voice may belong to a different account."
                  },
                  "source_upload": {
                    "type": "boolean",
                    "description": "Internal routing marker used by the local client to fail closed when an uploaded source has no affinity."
                  },
                  "tags": {
                    "type": "string",
                    "description": "Cover style and genre tags."
                  },
                  "title": {
                    "type": "string"
                  },
                  "prompt": {
                    "type": "string",
                    "description": "Lyrics; lyrics takes precedence when both fields are supplied."
                  },
                  "lyrics": {
                    "type": "string"
                  },
                  "negative_tags": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string",
                    "default": "suno-music"
                  },
                  "voice_id": {
                    "type": "string",
                    "description": "Optional public Suno +Voice UUID or https://suno.com/voice/<id> URL. Private Voices are rejected."
                  },
                  "persona_id": {
                    "type": "string",
                    "deprecated": true,
                    "description": "Legacy persona id. It is treated as a Voice when persona_model is voice_persona."
                  },
                  "persona_model": {
                    "type": "string",
                    "deprecated": true,
                    "enum": [
                      "style_persona",
                      "voice_persona"
                    ]
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "default": false
                  },
                  "pool": {
                    "type": "string",
                    "enum": [
                      "basic",
                      "super",
                      "heavy"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two submitted or completed Cover clips.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/audio_info"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing clip id or invalid account-affinity token."
          },
          "409": {
            "description": "The source-owning account is temporarily unavailable."
          },
          "429": {
            "$ref": "#/components/responses/ConcurrencyLimitExceeded"
          }
        }
      }
    },
...
    "/api/generate_stems": {
      "post": {
        "summary": "Make stem tracks (separate audio and music track).",
        "description": "Make stem tracks. Suno enforces clip ownership even when a clip is public/portable; pass the owning account affinity.",
        "tags": [
          "default"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "audio_id"
                ],
                "properties": {
                  "audio_id": {
                    "type": "string",
                    "description": "The ID of the song to generate stems for.",
                    "example": "e76498dc-6ab4-4a10-a19f-8a095790e28d"
                  },
                  "account_affinity": {
                    "type": "string",
                    "description": "Signed affinity for the account that owns the clip. Cross-account public/portable Stems are not supported upstream."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/schemas/audio_info"
          },
          "403": {
            "description": "Suno rejected Stems because the selected account does not own the clip."
          },
          "429": {
            "$ref": "#/components/responses/ConcurrencyLimitExceeded"
          }
        }
      }
    },
...
    "/api/get": {
      "get": {
        "summary": "Get audio information",
        "description": "",
        "tags": [
          "default"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "description": "Audio IDs, separated by commas. Leave blank to return a list of all music.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "account_affinity",
            "description": "Optional target-account affinity used for deterministic polling in multi-account mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "0",
                      "1"
                    ],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio_info"
                      },
                      {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/upload_audio/reconcile": {
      "post": {
        "summary": "Read-only reconciliation for a Studio Fast Upload operation.",
        "description": "Implemented by src/app/api/upload_audio/reconcile/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "post_api_upload_audio_reconcile",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "required": [
          "id",
          "title",
          "image_url",
          "lyric",
          "audio_url",
          "video_url",
          "created_at",
          "model_name",
          "status",
          "gpt_description_prompt",
          "prompt",
          "type",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "audio id"
          },
          "title": {
            "type": "string",
            "description": "music title"
          },
          "image_url": {
            "type": "string",
            "description": "music cover image"
          },
          "lyric": {
            "type": "string",
            "description": "music lyric"
          },
          "audio_url": {
            "type": "string",
            "description": "music download url"
          },
          "video_url": {
            "type": "string",
            "description": "Music video download link, can be used to share"
          },
          "created_at": {
            "type": "string",
            "description": "Create time"
          },
          "model_name": {
            "type": "string",
            "description": "Raw Suno provider model identifier used by the generated audio, for example chirp-fenix."
          },
          "status": {
            "type": "string",
            "description": "The generated states include submitted, queue, streaming, complete."
          },
          "gpt_description_prompt": {
            "type": "string",
            "description": "Simple mode on user input prompt, Suno will generate formal prompts, lyrics, etc."
          },
          "prompt": {
            "type": "string",
            "description": "The final prompt for executing the generation task, customized by the user in custom mode, automatically generated by Suno in simple mode."
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "tags": {
            "type": "string",
            "description": "Music genre. User-provided in custom mode, automatically generated by Suno in simple mode."
          }
        },
        "title": "audio_info",
        "description": "Audio Info"
      },
...
      "concurrency_limit_error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "message",
              "type",
              "code",
              "limit",
              "active_requests",
              "retry_after",
              "submission_state"
            ],
            "properties": {
              "message": {
                "type": "string",
                "example": "Global generation concurrency limit exceeded."
              },
              "type": {
                "type": "string",
                "enum": [
                  "rate_limit_error"
                ]
              },
              "code": {
                "type": "string",
                "enum": [
                  "concurrency_limit_exceeded"
                ]
              },
              "limit": {
                "type": "integer",
                "example": 4
              },
              "active_requests": {
                "type": "integer",
                "example": 4
              },
              "retry_after": {
                "type": "integer",
                "example": 5
              },
              "submission_state": {
                "type": "string",
                "enum": [
                  "not_submitted"
                ]
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "ConcurrencyLimitExceeded": {
        "description": "The global generation concurrency limit is full. Retry after a slot is released.",
        "headers": {
          "Retry-After": {
            "description": "Suggested retry delay in seconds.",
            "schema": {
              "type": "integer",
              "example": 5
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/concurrency_limit_error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "suno_admin_session"
      }
    }
  },
...
}